requirements.txt
httpx[http2]
kafka-python
websocket-client
prometheus-client
uvicorn[standard]
fastapi
